Move default certificate code to a separate file
* pkg/operator/controller/controller.go
(ensureDefaultCertificateForIngress, ensureDefaultCertificateDeleted): Move
from here...
* pkg/operator/controller/controller_default_certificate.go: ...to this new
file.
